pyranha: make the exception handling uniform across the code, and a potential bugfix in the real converter.
Now we never throw C++ exceptions directly, we use the Python api to set the Python exception and then raise with throw_already_set().

The bugfix consists of clearing the Python global error status in a place where we catch and ignore exceptions potentially originating form Python. If we do not do this, the error
status might be queried later by a function and generate an error even if no actual error has happened.