series: rework slightly series_is_rebindable so that it never results in hard assertion errors.
This makes it usable in contexts where it's preferrable to disable a template specialisation rather than erroring out,
e.g., in the series arithmetic operators. Previously, if a coefficient operator produced a type which is not a coefficient,
a hard compilation error was generated during the instantiation of the series operator. Now, the series operator
is simply disabled.