poly: introduce a global setting for selecting the GCD algorithm.
This is used mainly to force the use of a specific algorithm in case the algorithm recursively calls gcd(). Previously, if one specified PRS as
GCD algorithm, PRS would be used only at the first level of the recursion, as the recursion would internallly call gcd() without special
arguments (hence the heuristic GCD algorithm would be used eventually).